Refatorado código
- Criando os componentes
-- Card.js
-- InputCep.js
